==> 1 Process Zonas Eleitorais_RS.ipynb <==
# READ ME

This file process each city per run. You need first to find the URL of the webpage that generates the list of locais by navigating into the TRE-RS website and selecting the city. Then copy the URL and paste it in the "Download file from TRE-RS. This repository contains the HTML file for Porto Alegre.

As of today October 29th, one day before the second run, the website is returning an error code...

## Import libraries

from bs4 import BeautifulSoup
import pandas as pd
import urllib.request
import re

### Download file from tre-rs

Retrieve file from TRE-RS server.  When the website is functional, you should navigate there and selec the city you want to decode, and then copy the URL here.

url_to_decode = "https://eleitor.tre-rs.jus.br/locais-votacao/locais?q%5Bzona_numero_eq%5D=&q%5Bmunicipio_id_eq%5D=7994&button="
city_name = 'Porto Alegre'
# urllib.request.urlretrieve(url_to_decode, f"HTML files/Locais{city_name}2022.html")


### Import HTML into BeautifulSoup

html_soup = BeautifulSoup(open('HTML files/LocaisPortoAlegre2022.html'), "html.parser")

#### Create a dictionary with the Zonas Eleitorais and their position in the HTML code so the Secoes can be tied to the Zonas

dict_zonas = {}
div_content = html_soup.findAll(id= 'content')    
for div in div_content:
    for zona in div.find_all("h4"):
        zona_name = zona.text
        zona_line = zona.sourceline
        dict_zonas[zona_name] = zona_line

#### Create a DataFrame to be populated with the Zonas and Secoes information

secao_nome,lst_secoes,df_zonas,secao_endereco = "","","",""
columns = ['Zona', 'Local', 'Endereco', 'Secoes','Eleitores','Cidade']
df_zonas = pd.DataFrame(columns=columns)
div_content = html_soup.findAll(id= 'content')    
for div in div_content:
    for table in div.find_all("table"):
        for tr in table.find_all("tr"):
            td_c = 0
            for td in tr.findAll("td"):
                td_c+=1
                if td_c ==1:
                    # This is the first cell of the Secoes table, it contains the name and address of the secao place. 
                    # The information is split by a new line ('\n'), so I split by new line
                    secao_nome_endereco_lst = td.text.split("\n")
                    # Both name and address (positions 1 and 2 in the list) has extra blank characters, 
                    # used strip to clean the variable
                    secao_nome = secao_nome_endereco_lst[1].strip()
                    secao_endereco = secao_nome_endereco_lst[2].strip()
                    html_pos = td.sourceline
                    for i in dict_zonas:
                        if html_pos >= dict_zonas[i]:
                            numero_zona = i
                elif td_c == 2:
                    lst_secoes = []
                    #The second cell contains the list of secoes
                    secoes = td.find_all('span')
                    for s in secoes:
                        #print(s.text)
                        lst_secoes.append(s.text)
                elif td_c == 3:
                    secao_eleitores = td.text
                elif td_c == 4:
                    secao_cidade = td.text
            # print(secao_nome,lst_secoes)
            if secao_nome:
                df_zonas.loc[len(df_zonas.index)] = [numero_zona, secao_nome, secao_endereco, lst_secoes,secao_eleitores,secao_cidade]



df_zonas

def split_zonas(txt):
    return re.findall(r'\d+', txt)[0]

df_zonas['Zona_int'] = df_zonas['Zona'].apply(split_zonas)
df_zonas

df_zonas.to_csv("Output/Zonas and Secoes Porto Alegre.csv")

df_zonas_norm = pd.DataFrame()
columns = ['Zona', 'Local', 'Endereco', 'Secoes','Eleitores','Cidade']
df_zonas_norm = pd.DataFrame(columns=columns)
for index, row in df_zonas.iterrows():
    for i in row['Secoes']:
        df_zonas_norm.loc[len(df_zonas_norm.index)] = [row['Zona_int'],row['Local'],row['Endereco'],i,row['Eleitores'],row['Cidade']]
df_zonas_norm

df_zonas_norm

df_zonas_norm.to_csv("Output/Zonas and Secoes Porto Alegre NORMALIZED.csv")

### Backup - Old deprecated code that served as inspiration for the final version

It is impossible to reach good results without failing a lot. Here are my failed attempts:

# columns = ['Zona', 'Local', 'Endereco', 'Secoes','Eleitores','Cidade']
# df_zonas = ""
# df_zonas = pd.DataFrame(columns=columns)
# i=0
# lst_secoes = []
# # for tr in html_soup.select('table'):
# #     i+=1
# #     j=0

# for str in tr.select('tr'):
#     k=0
#     for td in str.select('td'):
#         k+=1
#         #print(td)
#         # print(k)
#         if k ==1:
#             local_secao, endereco_secao = "",""
#             # This is the first cell of the Secoes table, it contains the name and address of the secao place. 
#             # The information is split by a new line ('\n'), so I split by new line
#             secao_nome_endereco_lst = td.text.split("\n")
#             # Both name and address (positions 1 and 2 in the list) has extra blank characters, 
#             # used strip to clean the variable
#             secao_nome = secao_nome_endereco_lst[1].strip()
#             secao_endereco = secao_nome_endereco_lst[2].strip()
#             print(secao_nome)
#         elif k == 2:
#             lst_secoes = []
#             #The second cell contains the list of secoes
#             secoes = td.find_all('span')
#             for s in secoes:
#                 #print(s.text)
#                 lst_secoes.append(s.text)
#     df_zonas.loc[len(df_zonas.index)] = ['Zona', secao_nome, secao_endereco, lst_secoes,'Eleitores','Cidade']

# def create_df_1():
#     columns = ['Zona', 'Local', 'Endereco', 'Secoes','Eleitores','Cidade']
#     df_zonas = pd.DataFrame(columns=columns)
#     for tr in html_soup.select('table'):
#         i=0
#         for str in tr.select('tr'):
#             local_secao, endereco_secao, lst_secoes = "","",""
#             for line in str:
#                 #print(len(line),line)
#                     # print(line.select_one('span'))
#                 this_line = line.text.replace('\n', '|').split('|')
#                 int_this_l = int(len(this_line))
#                 #print(int_this_l,line.text)
#                 if int_this_l == 4 and len(this_line[1]) > 3:
#                     local_secao = this_line[1][12:]
#                     # print(local_secao)
#                     endereco_secao = this_line[2][14:]
#                     # print(endereco_secao)
#                 elif (int_this_l == 3 and len(this_line[1]) < 4):
#                     secoes = line.find('span')
#                     if secoes != -1 and secoes is not None:
#                         lst_secoes = []
#                         # print(secoes.text)
#                         lst_secoes.append(secoes.text)
#                 elif (int_this_l >= 4):
#                     secoes = line.find_all('span')
#                     lst_secoes = []
#                     for s in secoes:
#                         # print(s.text)
#                         lst_secoes.append(s.text)
#                 elif (int_this_l == 1):
#                     pass
#                     #print(line.text)
#             if len(local_secao) > 0:
#                 df_zonas.loc[len(df_zonas.index)] = ['Zona', local_secao, endereco_secao, lst_secoes,'Eleitores','Cidade']
#     return(df_zonas)
