==> 1 Process Zonas Eleitorais_SP.ipynb <==
# READ ME

Initially my intention was to analyze the votes in the city of Porto Alegre only. But later I decided to extend the analysis to the city of Sao Paulo, SP given the importance of this city for the country.

## Import libraries

import re
import pandas as pd

### Download file from tre-rs

1) [Click here to access the website with the link Locais de Votacao in SP](https://apps.tre-sp.jus.br/LocalVotacao/)
2) Click in "Clique aqui para fazer o download de todos locais de votação do Estado de São Paulo, em formato CSV (Excel)"
3) Move the file arquivo.csv into the CSV folder on this repository

### Read the file

df_locais_sp = pd.read_csv("CSV/arquivo.csv", encoding="Latin 1", header=None)
df_locais_sp.columns = ['Zona','Nome','Endereco','Cidade','Zonas','ZonasEspeciais']

### Create functions to find secoes using REGEX and function to append list of secoes especiais

def split_zonas(txt):
    return re.findall(r'\d+', txt)[0]

def split_secoes(txt):
    return re.findall(r'\d+', txt)

def apnd_zonas_e(lst1, lst2):
    return lst1.append(lst2)


df_locais_sp = df_locais_sp[df_locais_sp['Cidade']=='SÃO PAULO']

### Apply functions into the dataframe

df_locais_sp['Zona_int'] = df_locais_sp['Zona'].apply(split_zonas)
df_locais_sp['Secoes_int'] = df_locais_sp['Zonas'].apply(split_secoes)
df_locais_sp['Secoes_esp_int'] = df_locais_sp['ZonasEspeciais'].astype(str).apply(split_secoes)
df_locais_sp['Secoes_agg'] = df_locais_sp['Secoes_int'] + df_locais_sp['Secoes_esp_int']
df_locais_sp['Secoes_agg'] = df_locais_sp['Secoes_agg'].apply(set).apply(list)


df_locais_sp

### Export output to output folder

df_locais_sp.to_csv("Output/Zonas and Secoes Sao Paulo.csv")

df_zonas_norm = pd.DataFrame()
columns = ['Zona', 'Local', 'Endereco', 'Secoes','Cidade']
df_zonas_norm = pd.DataFrame(columns=columns)
for index, row in df_locais_sp.iterrows():
    for i in row['Secoes_agg']:
        df_zonas_norm.loc[len(df_zonas_norm.index)] = [row['Zona_int'],row['Nome'],row['Endereco'],i,row['Cidade']]
df_zonas_norm

df_zonas_norm.to_csv("Output/Zonas and Secoes Sao Paulo NORMALIZED.csv")
