==> 2 Process Zonas Eleitorais Secoes Votos.ipynb <==
## Import libraries

import pandas as pd
import urllib.request
import zipfile


### Download the results file from the TSE Dados Abertos

Uncomment when running the code for the first time, then you can comment again


# url_download = 'https://cdn.tse.jus.br/estatistica/sead/odsele/votacao_secao/votacao_secao_2022_BR.zip'

# opener = urllib.request.build_opener()
# opener.addheaders = [('User-agent', 'Mozilla/5.0')]
# urllib.request.install_opener(opener)
# urllib.request.urlretrieve(url_download, f"CSV/votacao_secao_2022_BR.zip")

If you prefer you can download the files manually: The full list of files can be found [here](https://dadosabertos.tse.jus.br/dataset/resultados-2022)!

### Unzip the downloaded file


Uncomment when running the code for the first time, then you can comment again

# with zipfile.ZipFile(f"CSV/votacao_secao_2022_BR.zip", 'r') as zip_ref:
#     zip_ref.extractall("CSV/")

### Read the extracted CSV

df = pd.read_csv("CSV/votacao_secao_2022_BR.csv", encoding="Latin 1", sep=";", usecols=['SG_UF','NM_MUNICIPIO','NR_ZONA','NR_SECAO','DS_CARGO','NM_VOTAVEL','QT_VOTOS'])

### Quick check of a specific Secao, Zona, City and State

First define a state, city, zona and secao

uf = 'SP'
city = 'SÃO PAULO'
zona = 5
secao = 62

check_secao = df[(df['NR_ZONA']==zona)&(df['NR_SECAO']==secao)&(df['SG_UF']==uf)&(df['NM_MUNICIPIO']==city)][['NM_VOTAVEL','QT_VOTOS']].groupby(by='NM_VOTAVEL').sum("QT_VOTOS")
check_secao

### Extractor of votes by Zona and Secao for a specific municipality and aggregate them by zone and secao

Need to define city and state and then run the code

uf = 'SP'
city = 'SÃO PAULO'

df_filtered = df[(df['SG_UF']==uf)&(df['NM_MUNICIPIO']==city)]
df_filtered = df_filtered[(df_filtered['NM_VOTAVEL']!='VOTO NULO')&(df_filtered['NM_VOTAVEL']!='VOTO BRANCO')]
df_results = df_filtered[['SG_UF','NM_MUNICIPIO','NR_ZONA','NR_SECAO']].drop_duplicates()
df_results['Lula_abslt'] = ""
df_results['Bolsonaro_abslt'] = ""
df_results['Lula_pctg'] = ""
df_results['Bolsonaro_pctg'] = ""
df_results['VotosTotais'] = ""

for i,r in df_results.iterrows():
    #print(r['NR_ZONA'],r['NR_SECAO'])
    zona = r['NR_ZONA']
    secao = r['NR_SECAO']
    results = df_filtered[(df_filtered['NR_ZONA']==zona)&(df_filtered['NR_SECAO']==secao)][['NM_VOTAVEL','QT_VOTOS']].groupby(by='NM_VOTAVEL').sum("QT_VOTOS")
    results['pctg'] = (results.QT_VOTOS / results.QT_VOTOS.sum())*100
    df_results.at[i,'Lula_pctg'] = results[results.index=='LUIZ INÁCIO LULA DA SILVA']['pctg'][0]
    df_results.at[i,'Bolsonaro_pctg'] = results[results.index=='JAIR MESSIAS BOLSONARO']['pctg'][0]
    df_results.at[i,'Lula_abslt'] = results[results.index=='LUIZ INÁCIO LULA DA SILVA']['QT_VOTOS'][0]
    df_results.at[i,'Bolsonaro_abslt'] = results[results.index=='JAIR MESSIAS BOLSONARO']['QT_VOTOS'][0]
    df_results.at[i,'VotosTotais'] = results.QT_VOTOS.sum()



df_results

df_filtered['QT_VOTOS'].sum()

df_results['VotosTotais'].sum()

df_presidente = df[df['DS_CARGO']=='PRESIDENTE']
df_presidente

# df_presidente[df_presidente['SG_UF']=='SP'].groupby(by=['NM_MUNICIPIO','NM_VOTAVEL']).sum("QT_VOTOS").to_clipboard()

df_filtered['NM_VOTAVEL'].unique()

### Import Zonas Eleitorais with addresses

Import the file corresponding to the city you are analyzing

df_zonas_add = pd.read_csv("Output/Zonas and Secoes Sao Paulo NORMALIZED.csv")

df_zonas_add

vote_counts_address = pd.merge(df_zonas_add, df_results[['NR_ZONA','NR_SECAO','Lula_abslt','Bolsonaro_abslt','Lula_pctg','Bolsonaro_pctg','VotosTotais']],  how='right', left_on=['Zona','Secoes'], right_on = ['NR_ZONA','NR_SECAO'])

vote_counts_address.drop(columns=['Unnamed: 0'], inplace=True)

vote_counts_address

vote_counts_address["VotosTotais"].sum()

## Possible analysis:

### 1) By name of zona

# local = 'EEEF PROF. LANGENDONCK'
# apuracao = vote_counts_address[(vote_counts_address['Local']==local)]
# votos_lula = apuracao['Lula_abslt'].sum()
# votos_bolsonaro = apuracao['Bolsonaro_abslt'].sum()
# votos_totais = apuracao['VotosTotais'].sum()
# print(f"\nNome Secao : {local} ")
# print(f"Votos totais na secao : {votos_totais} \n")
# print(f"Votos totais Lula na secao : {votos_lula} | Percentual: {(votos_lula/votos_totais)*100}%")
# print(f"Votos totais Bolsonaro na secao : {votos_bolsonaro} | Percentual: {(votos_bolsonaro/votos_totais)*100}%")
# print(f"Votos totais OUTROS (outros candidatos/branco/nulo) na secao : {votos_totais-votos_bolsonaro-votos_lula} | Percentual: {((votos_totais-votos_bolsonaro-votos_lula)/votos_totais)*100}%")


### 2) Places where Bolsonaro had less votes than 11%

bolsonaro_less_11pct = vote_counts_address[vote_counts_address['Bolsonaro_pctg']<11]
contagem = bolsonaro_less_11pct['Local'].count()
print(f"Em quantos locais bolsonaro teve menos de 11% dos votos: {contagem}")
print(f"Total de locais de votacao: {vote_counts_address['Local'].count()}")
print(f"Percentual de locais de votacao on Bolsonaro teve menos de 11%: {(contagem/vote_counts_address['Local'].count())*100}%")


print("Locais onde Bolsonaro teve menos de 11% dos votos:\n")
for i in bolsonaro_less_11pct['Local']:
    print(i)

vote_counts_address.to_excel("Output/VotesSaoPaulo.xlsx")
