==> 2 Votes NE.ipynb <==
## Import libraries

import pandas as pd
import urllib.request
import zipfile


### Download the results file from the TSE Dados Abertos

Uncomment when running the code for the first time, then you can comment again

# url_download = 'https://cdn.tse.jus.br/estatistica/sead/odsele/votacao_secao/votacao_secao_2022_BR.zip'
# opener = urllib.request.build_opener()
# opener.addheaders = [('User-agent', 'Mozilla/5.0')]
# urllib.request.install_opener(opener)
# urllib.request.urlretrieve(url_download, f"CSV/votacao_secao_2022_BR.zip")

If you prefer you can download the files manually: The full list of files can be found [here](https://dadosabertos.tse.jus.br/dataset/resultados-2022)!

### Unzip the downloaded file


Uncomment when running the code for the first time, then you can comment again

# with zipfile.ZipFile(f"CSV/votacao_secao_2022_BR.zip", 'r') as zip_ref:
#     zip_ref.extractall("CSV/")

### Read the extracted CSV

df = pd.read_csv("CSV/votacao_secao_2022_BR.csv", encoding="Latin 1", sep=";", usecols=['NR_TURNO','SG_UF','NM_MUNICIPIO','NR_ZONA','NR_SECAO','DS_CARGO','NM_VOTAVEL','QT_VOTOS'])

list_states = df['SG_UF'].unique()
# df['SG_UF'].unique()
df_ne = df[df['SG_UF'].isin(['BA','MA','PI','RN','PB','SE','CE','PE','AL'])]

# Quality check, total votes for Lula in Bahia
# df[(df['SG_UF']=='BA')&(df['NM_VOTAVEL']=='LUIZ INÁCIO LULA DA SILVA')&(df['DS_CARGO']=='PRESIDENTE')&(df['NR_TURNO']==2)]['QT_VOTOS'].sum()

### Filter by 2o turno and exclude null and branco votes, 

df_ne = df_ne[(df_ne['NR_TURNO']==2)].copy()

## RUN PIVOT

results_ne = df_ne[['SG_UF','NM_MUNICIPIO','NM_VOTAVEL','QT_VOTOS']].groupby(by=['SG_UF','NM_MUNICIPIO','NM_VOTAVEL']).sum("QT_VOTOS")
pivot_ne = results_ne.unstack(level='NM_VOTAVEL')
pivot_ne

results_ne_sessao = df_ne[['SG_UF','NM_MUNICIPIO','NR_ZONA','NR_SECAO','NM_VOTAVEL','QT_VOTOS']].groupby(by=['SG_UF','NM_MUNICIPIO','NR_ZONA','NR_SECAO','NM_VOTAVEL']).sum("QT_VOTOS")
pivot_ne_sessao = results_ne_sessao.unstack(level='NM_VOTAVEL')
pivot_ne_sessao

## Export PIVOT

pivot_ne_sessao.to_excel("Output/pivot_ne_sessao_2turno.xlsx", merge_cells=False)

# df_results = df_ne[['SG_UF','NM_MUNICIPIO']].copy()
# df_results['Lula_abslt'] = ""
# df_results['Bolsonaro_abslt'] = ""
# df_results['Lula_pctg'] = ""
# df_results['Bolsonaro_pctg'] = ""
# df_results['VotosTotais'] = ""
# df_results.drop_duplicates(inplace=True)
# df_results = df_results.sort_values(by='SG_UF')
# df_results

# for i,r in df_results.iterrows():
#     print(r['SG_UF'],r['NM_MUNICIPIO'])
#     uf = r['SG_UF']
#     muni = r['NM_MUNICIPIO']
#     results = df_ne[(df_ne['SG_UF']==uf)&(df_ne['NM_MUNICIPIO']==muni)][['NM_VOTAVEL','QT_VOTOS']].groupby(by='NM_VOTAVEL').sum("QT_VOTOS")
#     results['pctg'] = (results.QT_VOTOS / results.QT_VOTOS.sum())*100
#     df_results.at[i,'Lula_pctg'] = results[results.index=='LUIZ INÁCIO LULA DA SILVA']['pctg'][0]
#     df_results.at[i,'Lula_abslt'] = results[results.index=='LUIZ INÁCIO LULA DA SILVA']['QT_VOTOS'][0]
#     bolsonaro_pctg = results[results.index=='JAIR MESSIAS BOLSONARO']['pctg'][0]
#     bolsonaro_abslt = results[results.index=='JAIR MESSIAS BOLSONARO']['QT_VOTOS'][0]    
#     df_results.at[i,'Bolsonaro_pctg'] = bolsonaro_pctg
#     df_results.at[i,'Bolsonaro_abslt'] = bolsonaro_abslt
#     df_results.at[i,'VotosTotais'] = results.QT_VOTOS.sum()

# df_results.sort_values(by='Bolsonaro_abslt')

### Check specific Sessao and Zona

zona = 195
secao = 76
check_secao = df_ne[(df_ne['NR_ZONA']==zona)&(df_ne['NR_SECAO']==secao)][['NM_VOTAVEL','QT_VOTOS']].groupby(by='NM_VOTAVEL').sum("QT_VOTOS")
df_ne[(df_ne['NR_ZONA']==zona)&(df_ne['NR_SECAO']==secao)][['SG_UF','NM_MUNICIPIO','NM_VOTAVEL','QT_VOTOS']]
