==> 5 Analise Municipais 2020.ipynb <==
import pandas as pd
import geopandas as gpd
import matplotlib.pyplot as plt
import contextily as ctx
from shapely.geometry import Point
import plotly_express as px

### Municipal votes count for a specific vereador of Porto Alegre

df = pd.read_csv(r"CSV/votacao_secao_2020_RS.csv", encoding="Latin 1", sep=";")
df.columns

#### Import geocoded places

geocoded_df = pd.read_csv("CSV/geocoded_voting_places_001.csv")
df_lat_long = df.merge(geocoded_df, on = ['CD_MUNICIPIO', 'NR_ZONA', 'NR_LOCAL_VOTACAO'], how = 'left')
df_lat_long.head()

#### Import nomes de locais

nomeslocais = pd.read_csv("CSV/NomesLocais.csv")
nomeslocais.head()

### Merge geocoded and nome locais

df_votos_vereador_nmlocal = pd.merge(df_lat_long, nomeslocais, on = ['NM_MUNICIPIO','NR_ZONA','NR_SECAO'])
df_votos_vereador_nmlocal.shape

#### Check for Nome do Vereador

df_votos_vereador = df_votos_vereador_nmlocal[df_votos_vereador_nmlocal['NM_VOTAVEL']=='LUCAS SALOMÓN DA SILVA FUHR'][['NM_MUNICIPIO','NM_VOTAVEL','NM_LOCAL_VOTACAO', 'NR_ZONA','QT_VOTOS']]

#### Dataquality check

df_votos_vereador['QT_VOTOS'].sum()

#### Export

df_votos_vereador_g = df_votos_vereador[['NR_ZONA','NM_LOCAL_VOTACAO','QT_VOTOS']].groupby(by=['NR_ZONA','NM_LOCAL_VOTACAO']).sum('QT_VOTOS')
df_votos_vereador_g.to_excel('votos_fuhr_2020.xlsx', merge_cells=False)
